+++ src/main/java/gov/usgs/cida/wqp/swagger/SwaggerConfig.java
@@ -50,7 +50,7 @@ public class SwaggerConfig {
 	public static final String SIMPLE_STATION_TAG_NAME = "Simple Station (deprecated)";
 	public static final String STATION_TAG_NAME = "Station";
 	public static final String SUMMARY_ORGANIZATION_TAG_NAME = "Summary Organizaton";
-	public static final String SUMMARY_STATION_TAG_NAME = "Summary Monitoring Location";
+	public static final String SUMMARY_MONITORING_LOCATION_TAG_NAME = "Summary Monitoring Location";
 	public static final String VERSION_TAG_NAME = "Application Version";
 	public static final String VERSION_TAG_DESCRIPTION = "Display Application Version";
 	public static final String PROJECT_TAG_NAME = "Project";
@@ -103,7 +103,7 @@ public class SwaggerConfig {
 					new Tag(SIMPLE_STATION_TAG_NAME, TAG_DESCRIPTION),
 					new Tag(STATION_TAG_NAME, TAG_DESCRIPTION),
 					new Tag(SUMMARY_ORGANIZATION_TAG_NAME, TAG_DESCRIPTION),
-					new Tag(SUMMARY_STATION_TAG_NAME, TAG_DESCRIPTION),
+					new Tag(SUMMARY_MONITORING_LOCATION_TAG_NAME, TAG_DESCRIPTION),
 					new Tag(VERSION_TAG_NAME, VERSION_TAG_DESCRIPTION),
 					new Tag(FILE_DOWNLOAD_TAG_NAME, FILE_DOWNLOAD_TAG_NAME)
 				)