+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Station summary service returns only sites that have summary information.
 - Fixed ActivityFileUrl, ActivityMetricURL, and ActivityAttachedBinaryObject logic errors.
 - Refactored Summary Station mapper to use new format of station_sum table.
+- Changed use of 'summarystation' to summaryMonitoringLocation.' 
 
 ## [0.15.0]
 ### Added